+++ packages/stream_chat_flutter/lib/src/components/stream_chat_component_builders.dart
@@ -22,6 +22,7 @@ Iterable<StreamComponentBuilderExtension<Object>> streamChatComponentBuilders({
   StreamComponentBuilder<StreamMessageFooterProps>? messageFooter,
   StreamComponentBuilder<StreamMessageComposerAttachmentListProps>? messageComposerAttachmentList,
   StreamComponentBuilder<StreamMessageComposerAttachmentProps>? messageComposerAttachment,
+  StreamComponentBuilder<StreamMentionTileProps>? mentionTile,
 
   // ── Attachments ──────────────────────────────────────────────────────────
   StreamComponentBuilder<StreamImageAttachmentProps>? imageAttachment,
@@ -54,6 +55,7 @@ Iterable<StreamComponentBuilderExtension<Object>> streamChatComponentBuilders({
     if (messageFooter != null) StreamComponentBuilderExtension(builder: messageFooter),
     if (messageComposerAttachmentList != null) StreamComponentBuilderExtension(builder: messageComposerAttachmentList),
     if (messageComposerAttachment != null) StreamComponentBuilderExtension(builder: messageComposerAttachment),
+    if (mentionTile != null) StreamComponentBuilderExtension(builder: mentionTile),
     if (imageAttachment != null) StreamComponentBuilderExtension(builder: imageAttachment),
     if (videoAttachment != null) StreamComponentBuilderExtension(builder: videoAttachment),
     if (giphyAttachment != null) StreamComponentBuilderExtension(builder: giphyAttachment),