+++ packages/stream_chat_flutter/lib/stream_chat_flutter.dart
@@ -197,6 +197,7 @@ export 'src/media_gallery_preview/stream_media_gallery_preview_header.dart';
 export 'src/media_gallery_preview/stream_media_gallery_preview_item.dart';
 export 'src/media_gallery_preview/video_player/stream_video_player.dart';
 export 'src/mention/mention.dart';
+export 'src/mention/mention_tile.dart';
 export 'src/message_action/message_action.dart';
 export 'src/message_action/message_actions_builder.dart';
 export 'src/message_input/attachment_picker/stream_attachment_picker.dart';